Update README.md
Code for Code2040 api challenge. 
Delta time is for the last challenge, take a look at it, it does everything right but some how the math is off so answer is 2 hrs off when done forward and reverse way. Probably an issue with local time vs UTC , I do not know how to fix but this code gets in the ball park by about 2 hours every time often even the seconds are correct. 

I did all the code in python, I could do it in java, but I am not as fluid. I used functional programming and recursion, I can implement loop structures I simply thought it was simpler and clearer using recursion and functional programming